refactor: 다이얼로그 고도화
