+++ packages/ui/Dialog/DialogProvider.tsx
@@ -1,54 +1,47 @@
 'use client';
 
-import React, { ChangeEvent, createContext, useCallback, useEffect, useState } from 'react';
+import { createContext, useEffect, useState } from 'react';
 import { DialogComponent } from './DialogComponent';
 import { DialogOptionType, ProviderChildren } from './types';
 
 export const DialogContext = createContext({
   openDialog(option: DialogOptionType) {},
   closeDialog() {},
-  checkCheckBox(e: ChangeEvent<HTMLInputElement>) {},
+  checkCheckBox(isCheckValue:boolean) {},
 });
 
 function DialogProvider({ children }: ProviderChildren) {
   const [dialogOption, setDialogOption] = useState<DialogOptionType | null>(null);
-  const [isMounted, setIsMounted] = useState(false);
-  useEffect(() => {
-    setIsMounted(true);
-  });
+  const [isCheck, setIsCheck] = useState(dialogOption?.checkBoxOptions?.checked ?? false);
 
-  const openDialog = useCallback(
-    (option: DialogOptionType) => {
-      setDialogOption(option);
-    },
-    [dialogOption]
-  );
+  const openDialog = (option: DialogOptionType) => {
+    setDialogOption(option);
+  };
 
   const closeDialog = () => {
     setDialogOption(null);
   };
 
-  const checkCheckBox = (e: ChangeEvent<HTMLInputElement>) => {
-    setDialogOption((prevOption) => ({
-      ...prevOption,
-      checkBoxOptions: { ...prevOption?.checkBoxOptions, checked: e.target.checked },
-    }));
+  const checkCheckBox = (isCheckValue:boolean) => {
+    setIsCheck(isCheckValue);
   };
+
+  useEffect(() => {
+    if (dialogOption?.checkBoxOptions) {
+      const newCheckBoxOption = { ...dialogOption.checkBoxOptions, checked: isCheck };
+      const newDialogOption = { ...dialogOption, checkBoxOptions: newCheckBoxOption };
+      setDialogOption(newDialogOption);
+    }
+  }, [isCheck]);
 
   return (
     <>
-      {isMounted && (
-        <DialogContext.Provider value={{ openDialog, closeDialog, checkCheckBox }}>
-          {children}
-          {dialogOption && (
-            <DialogComponent
-              isOpen={dialogOption !== null}
-              onClose={closeDialog}
-              {...dialogOption}
-            />
-          )}
-        </DialogContext.Provider>
-      )}
+      <DialogContext.Provider value={{ openDialog, closeDialog, checkCheckBox }}>
+        {children}
+        {dialogOption && (
+          <DialogComponent isOpen={dialogOption !== null} onClose={closeDialog} {...dialogOption} />
+        )}
+      </DialogContext.Provider>
     </>
   );
 }
