+++ packages/ui/index.ts
@@ -1,7 +1,9 @@
 export * from "./cssVariables";
 
 // component exports
-// export { default as Button } from "./Button";
-export { default as Dialog } from "./Dialog";
-// export { ToastProvider, useToast } from "./Toast";
-// export type { ToastOptionType } from "./Toast";
+export { default as Button } from "./Button";
+export { Dialog, DialogContext, DialogProvider, useDialog } from "./Dialog";
+export type { DialogOptionType } from "./Dialog";
+export { ToastProvider, useToast } from "./Toast";
+export type { ToastOptionType } from "./Toast";
+
